requirements.txt
evaluate
rouge-score
bert-score
pandas
matplotlib
docx2txt
pdfminer.six